Move rgb_thread way from __init__.py for clarity
